+++ protozfits/digicam.py
@@ -1,16 +1,42 @@
+from os.path import isfile
 import warnings
 import numpy as np
 from astropy.utils.decorators import lazyproperty
 
+from . import rawzfitsreader
 from google.protobuf.pyext.cpp_message import GeneratedProtocolMessageType
-from .simple import File
 from .patch_ids import PATCH_ID_INPUT_SORT_IDS, PATCH_ID_OUTPUT_SORT_IDS
 from .any_array_to_numpy import any_array_to_numpy
+from .L0_pb2 import CameraEvent
 
 
-def event_source(path, run_id=0):
-    for event in File(path).Events:
-        yield Event(event, run_id)
+class File:
+
+    def __init__(self, fname):
+        if not isfile(fname):
+            raise FileNotFoundError(fname)
+        self.fname = fname
+        rawzfitsreader.open(self.fname + ":Events")
+        self.numrows = rawzfitsreader.getNumRows()
+        self.run_id = 0
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        event = CameraEvent()
+        try:
+            event.ParseFromString(rawzfitsreader.readEvent())
+            return Event(event, self.run_id)
+        except EOFError:
+            raise StopIteration
+
+    def list_tables(self):
+        return rawzfitsreader.listAllTables(self.fname)
+
+    def rewind_table(self):
+        # Rewind the current reader. Go to the beginning of the table.
+        rawzfitsreader.rewindTable()
 
 
 class Event:
